feat: extract translate
